Move `RequestedVersion` to resolve.rs and implement `Display`
Also adds `println` to `huak_python_formatter` cli install operation
